algo(quickSort): format + IO changes
