--- src/algorithm_practice/Algorithms_From_Scratch/atoi/test.in
@@ -1,8 +1,0 @@
-15a: 
-denver2: 
-den22ver: 
-22denver: 
-15a: 
-denver2: 
-den22ver: 
-22denver:
