--- src/algorithm_practice/Algorithms_From_Scratch/atoi/test.out
@@ -1,8 +1,0 @@
-15 -- 15
-0 -- 0
-0 -- 0
-22 -- 22
-15 -- 15
-0 -- 0
-0 -- 0
-22 -- 22
