--- src/algorithm_practice/Searching_And_Sorting/quicksort/test.in
@@ -1,10 +1,0 @@
-23232
-343
-232
-235
-1
-132
-234
-2345
-345
-111
