--- src/algorithm_practice/Searching_And_Sorting/quicksort/test.out
@@ -1,10 +1,0 @@
-0
-1
-111
-132
-232
-234
-235
-343
-345
-2345
